Modifications for training on ISVQA like in ISVQA paper
I changes that ISVQA dataset class so it outputs two targets. One for calculating the BCELossWithLogits for Mulit-label loss where for each right answer 'regardless of frequency' the model learns gets a loss penalty for having too low probability for it. I also changed the accuracy function to do it like they did it (half a point if it matches one answer, a full point if it matches at least two answers). I made the necessary modifications on train.py so that it runs